Changes:-
added combat scene between Dog and Martian Animal.

Note for self:-

There is a room of improvement for the code. Overall, it works but the OOP paradigm is not fully enforced.

I admit there are redundancy here and there which can be refactored such as the movement of entities (can create a Movement class but this is my technical debt I guess since time constraint).

There are also bugs in the file that I have not debug yet.